Reorder the code so that auxiliary functions appear first in the .cpp files and are in an unnamed namespace.  Move Elk to elliptic_integrals.{hpp,cpp} because, well, it's an integral.
